bp edit

+++ src/factory/general/Part.java
@@ -86,7 +86,7 @@ public class Part implements Serializable {
 //        this.img = new ImageIcon(fn);
         //this.number=time()\;
 	this.guipart = null;
-	this.number = System.currentTimeMillis()/1000;
+	this.number = (int)(System.currentTimeMillis()/(long)1000);
     }
     public void setFilename(String in)
     {
